feat: add exercise stats chart data endpoint
god forgive me for those tests

+++ src/src/main/java/xyz/cursedman/gym_api/domain/dtos/progressStatistics/TimeSeriesDto.java
@@ -13,5 +13,5 @@ import java.util.List;
 @Builder
 public class TimeSeriesDto {
 	private List<String> labels;
-	private List<Integer> values;
+	private List<Float> values;
 }
